Initial attempt at "text-input" wayland protocol for IME
This is my first crack at allowing devdraw to receive text-input-v3
events, which allows it to accept text from input methods like ibus
and fcitx5.

I do not know of a great way to determine the location of the text
being input into an application, so for input methods which use a
pop-up to select from multiple candidates, that pop-up menu will
appear at a fixed location on the screen, usually the top-left corner
of the window, depending on your compositor.

Another issue is that the updates to the text as it is being composed
and committed are slow enough that they become visible. I have not
determined why that is the case.

However, this change works well enough for me to compose text in
Japanese, using fcitx-anthy, and Vietnamese, using fcitx-unikey.